feat: Add image downloading functionality
This commit adds the ability to download images using the app. It uses the fs module to save the images locally, axios for making HTTP requests, and cheerio for parsing HTML content. The image URLs are hardcoded for now, but will be dynamically fetched in future updates.

The commit also includes a function check_if_update(), which compares the current version of the app with the latest version available on GitHub. If an update is available, it will return true. This functionality will be used to notify users of new updates.

This commit should trigger a release bumping a MINOR version.