feat: Add InstallationScene component
This commit adds a new feature for the user in the form InstallationScene component. This component is responsible for displaying the installation progress to user. It includes a header, a console output section, and uses the useState and useEffect hooks from React.

The useEffect hook is used to send a command to the API and receive the output from it. The received output is then set in the console_output state using the setConsole_output function.

The component is styled using tailwind CSS classes.

This commit will trigger a release bumping a MINOR version.