feat: Create start file and set config
This commit adds the functionality to create a start file for the user and set the configuration values in the local file. The start file is created with the appropriate target, description, and icon. The configuration values are modified in the local file according to the user's input. The event sender is used to send the success or ok messages to the user.