feat: Add useEffect hook for checking if button should be disabled
This hook to the App component in order to check if the finish button should be disabled. If the activeStep is equal to the last step in the steps array, the toDisable state is set to false. This will allow the finish button to be enabled when all steps have been completed.