feat: Add createUser function for initializing config
This commit adds the createUser function that is used to create a user during the initialization of the config. The function makes a PUT request to the API endpoint to create a new user. If successful, the user ID is returned and stored in the config. If there is an error, appropriate error messages are displayed. This commit is a new feature for the user and will trigger aingOR version.